feature : 할일 문서에서 수정및 삭제 작성 기능 구현

+++ src/main/java/com/wegoing/dao/MemberDAO.java
@@ -7,7 +7,6 @@ import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
-import com.wegoing.dto.ClubMemberInfo;
 import com.wegoing.dto.MemberDTO;
 
 @Repository
@@ -19,5 +18,5 @@ public interface MemberDAO {
 	String findEmailByNameAndTel(MemberDTO dto); // 이메일 찾기
 	void updatePw(MemberDTO dto); // 비밀번호 재설정
 	List<MemberDTO> selectMyPartners(String email); // 내파트너 조회하기, 이미지도 불러와야해서 MemberDAO이용
-	List<MemberDTO> selectClubMembersInfo(long clno); 
+	List<MemberDTO> selectClubMembersInfo(int clno); 
 }
